meal_planner/requirements.txt
python-dotenv>=1.0
requests>=2.31
gspread>=6.0
Pillow>=10.0
